add test and rm old
