+++ tests/unit/conftest.py
@@ -45,6 +45,7 @@ from featurebyte.models.task import Task as TaskModel
 from featurebyte.models.tile import TileSpec
 from featurebyte.query_graph.graph import GlobalQueryGraph
 from featurebyte.routes.lazy_app_container import LazyAppContainer
+from featurebyte.routes.registry import app_container_config
 from featurebyte.schema.task import TaskStatus
 from featurebyte.schema.worker.task.base import BaseTaskPayload
 from featurebyte.service.task_manager import TaskManager
@@ -1365,6 +1366,7 @@ def app_container_fixture(persistent):
         task_manager=task_manager,
         storage=LocalTempStorage(),
         catalog_id=DEFAULT_CATALOG_ID,
+        app_container_config=app_container_config,
     )
